sphere switcher
created a sphere switching function. It keeps an array of spheres and a reference to the current sphere. It has a public function which takes an index as a variable which it can use to load any sphere from the array. Right now it has a coroutine to swap spheres after 3 seconds.